double doube toil and trouble, fire burns and warnings bubble
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@10492 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/c-ares/ares_search.c
@@ -236,7 +236,7 @@ static int single_domain(ares_channel channel, const char *name, char **s)
               while ((status = ares__read_line(fp, &line, &linesize))
                      == ARES_SUCCESS)
                 {
-                  if (strncasecmp(line, name, len) != 0 ||
+                  if (strncasecmp(line, name, (int)len) != 0 ||
                       !isspace((unsigned char)line[len]))
                     continue;
                   p = line + len;
